Rename package "resource" to "handler"
